+++ src/Vipentti.Nuke.Components/StandardPublishGitHubActionsAttribute.cs
@@ -1,5 +1,7 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
+using Nuke.Common.CI;
 using Nuke.Common.CI.GitHubActions;
 using Nuke.Common.CI.GitHubActions.Configuration;
 using Nuke.Common.Execution;
@@ -27,6 +29,22 @@ public class StandardPublishGitHubActionsAttribute : ExtendedGitHubActionsAttrib
         nameof(IPublishPackagesToNuGet.Publish),
     };
 
+    public static readonly GitHubActionsPermissions[] DefaultReadPermissions =
+        Array.Empty<GitHubActionsPermissions>();
+
+    public static readonly GitHubActionsPermissions[] DefaultWritePermissions = new[]
+    {
+        GitHubActionsPermissions.Actions,
+        GitHubActionsPermissions.Checks,
+        GitHubActionsPermissions.Contents,
+        GitHubActionsPermissions.Deployments,
+        GitHubActionsPermissions.Issues,
+        GitHubActionsPermissions.Discussions,
+        GitHubActionsPermissions.Packages,
+        GitHubActionsPermissions.Pages,
+        GitHubActionsPermissions.Statuses,
+    };
+
     public const string DefaultPublishCondition = "${{ runner.os == 'Windows' }}";
 
     public StandardPublishGitHubActionsAttribute(
@@ -36,10 +54,31 @@ public class StandardPublishGitHubActionsAttribute : ExtendedGitHubActionsAttrib
     )
         : base(name, image, images)
     {
+        WritePermissions = DefaultWritePermissions;
         PublishArtifacts = true;
         PublishCondition = DefaultPublishCondition;
         EmptyWorkflowTrigger = true;
         FetchDepth = 0;
+    }
+
+    public override ConfigurationEntity GetConfiguration(
+        IReadOnlyCollection<ExecutableTarget> relevantTargets
+    )
+    {
+        if (IncludeDefaultWritePermissions)
+        {
+            WritePermissions = WritePermissions
+                .Concat(DefaultWritePermissions)
+                .Distinct()
+                .ToArray();
+        }
+
+        if (IncludeDefaultReadPermissions)
+        {
+            ReadPermissions = ReadPermissions.Concat(DefaultReadPermissions).Distinct().ToArray();
+        }
+
+        return base.GetConfiguration(relevantTargets);
     }
 
     protected override GitHubActionsJob GetJobs(
@@ -65,6 +104,8 @@ public class StandardPublishGitHubActionsAttribute : ExtendedGitHubActionsAttrib
         return base.GetJobs(image, relevantTargets);
     }
 
+    public bool IncludeDefaultReadPermissions { get; set; } = true;
+    public bool IncludeDefaultWritePermissions { get; set; } = true;
     public bool IncludeDefaultImportSecrets { get; set; } = true;
     public bool IncludeDefaultImportVars { get; set; } = true;
     public bool IncludeDefaultInvokedTargets { get; set; } = true;