+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Enable default write permissions in StandardPublishGitHubActions
 
 ## [0.2.2] / 2023-11-09
 - Revert changes: When using ICreateGitHubRelease do not create tag manually in IFinalizeRelease